+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # Changelog
 
+## [1.6.5] - 2025-06-02
+
+### Changed
+
+- Fix the `deployTrigger` method so that it sends the workflow ID to the API
+
 ## [1.6.4] - 2025-05-30
 
 ### Added